Seperate the multer middleware - backend optimization
